+++ src/lib.rs
@@ -8,6 +8,7 @@ pub mod bit_set_n;
 pub mod bit_set_shiftable;
 pub mod bit_set_trait;
 mod n_choose_k;
+pub mod set_size_n_iter;
 
 pub type SetElement = u32;
 pub use bit_set_array::*;